Created new controller and make some changes on crud

+++ app/Http/Controllers/AccountController.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Http\Controllers\Auth;
+namespace App\Http\Controllers;
 
 use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
@@ -13,8 +13,8 @@ class AccountController extends Controller
     {
         $account = Account::findOrFail($id);
         $validationForAccount = Validator::make($request->all(), [
-            'account_name'   => 'required | max:40',
-            'account_number' => 'required | numeric | unique:accounts,account_number',
+            'account_name'   => 'required|max:40|alpha',
+            'account_number' => 'required|numeric|unique:accounts,account_number',
         ]);
 
         if ($validationForAccount->fails()) {
@@ -33,18 +33,28 @@ class AccountController extends Controller
 
     public function delete($id)
     {
+
         $account = Account::findOrFail($id);
 
-        if ($account->is_default == true) {
+        if ($account) {
+
+            if ($account->is_default == true) {
+                return response()->json([
+                    'status'    => false,
+                    'message'   => 'Deletation not allowed for default account'
+                ]);
+            } else {
+                $account->delete();
+                return response()->json([
+                    'status'    => true,
+                    'message'   => 'Account Deleted Successfully...'
+                ]);
+            }
+        }
+        else{
             return response()->json([
                 'status'    => false,
-                'message'   => 'Deletation not allowed for default account'
-            ]);
-        } else {
-            $account->delete();
-            return response()->json([
-                'status'    => true,
-                'message'   => 'Account Deleted Successfully...'
+                'message'   => 'No Account Found'
             ]);
         }
     }
@@ -60,9 +70,9 @@ class AccountController extends Controller
     public function insert(Request $request)
     {
         $validationForAccount = Validator::make($request->all(), [
-            'account_name'   => 'required | max:40',
-            'account_number' => 'required | numeric | unique:accounts,account_number',
-            'user_id'        => 'numeric | required | exists:users,id'
+            'account_name'   => 'required|max:40|alpha',
+            'account_number' => 'required|numeric|unique:accounts,account_number',
+            'user_id'        => 'numeric|required|exists:users,id'
         ]);
 
         if ($validationForAccount->fails()) {
@@ -73,22 +83,21 @@ class AccountController extends Controller
             ]);
         }
 
-        $account = Account::create($request->only(['account_name','account_number','user_id']));
+        $account = Account::create($request->only(['account_name', 'account_number', 'user_id']));
 
         return response()->json([
             'status'         => true,
             'message'        => 'Account Created Successfully',
             'data'           => $account
         ], 200);
-
-
     }
 
     public function list()
     {
+        $account = Account::all();
         return response()->json([
             'message'   => 'Data',
-            'data'      => Account::all()
+            'data'      => $account
         ]);
     }
 }
